+++ KeenClient/KeenClient.m
@@ -20,6 +20,7 @@ static BOOL authorizedGeoLocationAlways = NO;
 static BOOL authorizedGeoLocationWhenInUse = NO;
 static BOOL geoLocationEnabled = NO;
 static BOOL loggingEnabled = NO;
+static BOOL geoLocationRequestEnabled = YES;
 static KIODBStore *dbStore;
 
 @interface KeenClient ()
@@ -225,6 +226,16 @@ static KIODBStore *dbStore;
     KCLog(@"Disabling Geo Location");
     geoLocationEnabled = NO;
 }
++ (void)enableGeoLocationDefaultRequest {
+    KCLog(@"Enabling Geo Location Request");
+    geoLocationRequestEnabled = YES;
+}
+
++ (void)disableGeoLocationDefaultRequest {
+    KCLog(@"Disabling Geo Location Request");
+    geoLocationRequestEnabled = NO;
+}
+
 
 + (void)clearAllEvents {
     [dbStore deleteAllEvents];
@@ -356,17 +367,24 @@ static KIODBStore *dbStore;
     if (geoLocationEnabled == YES) {
         KCLog(@"Geo Location is enabled.");
         // set up the location manager
-        if (self.locationManager == nil) {
-            if ([CLLocationManager locationServicesEnabled]) {
-                self.locationManager = [[CLLocationManager alloc] init];
-                self.locationManager.delegate = self;
-            }
+        if (self.locationManager == nil && [CLLocationManager locationServicesEnabled]) {
+            self.locationManager = [[CLLocationManager alloc] init];
+            self.locationManager.delegate = self;
         }
-        
+      
         // check for iOS 8 and provide appropriate authorization for location services
+      
+        if(self.locationManager != nil)
+        {
+            // If location services are already authorized, then just start monitoring.
+            CLAuthorizationStatus clAuthStatus = [CLLocationManager authorizationStatus];
+            if ([KeenClient isLocationAuthorized:clAuthStatus]) {
+                [self startMonitoringLocation];
+            }
 #if __IPHONE_OS_VERSION_MAX_ALLOWED >= 80000
-        if(self.locationManager != nil) {
-            if([self.locationManager respondsToSelector:@selector(requestAlwaysAuthorization)]) {
+            // Else, try and request permission for that.
+            else if (geoLocationRequestEnabled &&
+                     [self.locationManager respondsToSelector:@selector(requestAlwaysAuthorization)]) {
                 // allow explicit control over the type of authorization
                 if(authorizedGeoLocationAlways) {
                     [self.locationManager requestAlwaysAuthorization];
@@ -379,17 +397,33 @@ static KIODBStore *dbStore;
                     [self.locationManager requestWhenInUseAuthorization];
                 }
             }
-        }
 #endif
-        
-        // if, at this point, the location manager is ready to go, we can start location services
-        if (self.locationManager) {
-            [self.locationManager startUpdatingLocation];
-            KCLog(@"Started location manager.");
         }
+      
     } else {
         KCLog(@"Geo Location is disabled.");
     }
+}
+
+-(void)startMonitoringLocation {
+    if(self.locationManager) {
+        [self.locationManager startUpdatingLocation];
+        KCLog(@"Started location manager.");
+    }
+}
+
++(BOOL)isLocationAuthorized:(CLAuthorizationStatus)status {
+#if TARGET_OS_IOS
+  if (status == kCLAuthorizationStatusAuthorizedWhenInUse ||
+      status == kCLAuthorizationStatusAuthorizedAlways) {
+    return YES;
+  }
+#elif TARGET_OS_MAC
+  if (status == kCLAuthorizationStatusAuthorized) {
+    return YES;
+  }
+#endif
+  return NO;
 }
 
 // Delegate method from the CLLocationManagerDelegate protocol.
@@ -410,6 +444,16 @@ static KIODBStore *dbStore;
     } else {
         KCLog(@"Event wasn't recent enough: %+.2d", (int)fabs(howRecent));
     }
+}
+
+-(void)locationManager:(CLLocationManager *)manager didChangeAuthorizationStatus:(CLAuthorizationStatus)status {
+    if ([KeenClient isLocationAuthorized:status]) {
+        [self startMonitoringLocation];
+    }
+}
+-(void)locationManager:(CLLocationManager *)manager didFailWithError:(NSError *)error
+{
+  KCLog(@"locationManager-didFailWithError: %@", [error localizedDescription]);
 }
 
 # pragma mark - Add events