+++ KeenClientExample/KeenClientExample/ThirdViewController.m
@@ -11,7 +11,7 @@
 
 @implementation ThirdViewController
 
-- (id)initWithNibName:(NSString *)nibNameOrNil bundle:(NSBundle *)nibBundleOrNil
+- (instancetype)initWithNibName:(NSString *)nibNameOrNil bundle:(NSBundle *)nibBundleOrNil
 {
     self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
     if (self) {
@@ -51,9 +51,9 @@
                                             error:nil];
         
         NSLog(@"response: %@", responseDictionary);
-        NSLog(@"error: %@", [error localizedDescription]);
+        NSLog(@"error: %@", error.localizedDescription);
         
-        NSNumber *result = [responseDictionary objectForKey:@"result"];
+        NSNumber *result = responseDictionary[@"result"];
         
         NSLog(@"result: %@", result);
         
@@ -61,10 +61,10 @@
         //NSNumber *resultValue = [[responseDictionary objectForKey:@"result"][0] objectForKey:@"result"];
         //NSLog(@"resultValue: %@", resultValue);
         
-        if(error || [responseDictionary objectForKey:@"error_code"]) {
-            self.resultTextView.text = [NSString stringWithFormat:@"Failure! 😞 \n\n error: %@\n\n response: %@", [error localizedDescription] ,[responseDictionary description]];
+        if(error || responseDictionary[@"error_code"]) {
+            self.resultTextView.text = [NSString stringWithFormat:@"Failure! 😞 \n\n error: %@\n\n response: %@", error.localizedDescription ,responseDictionary.description];
         } else {
-            self.resultTextView.text = [NSString stringWithFormat:@"Success! 😄 \n\n response: %@", [responseDictionary description]];
+            self.resultTextView.text = [NSString stringWithFormat:@"Success! 😄 \n\n response: %@", responseDictionary.description];
         }
     };