+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+### Added
+- Added support for using HTTP proxies. #171
+
 ### Changed
 - Updated Swift example project to Swift 3 syntax. #175
 - Added method to disable automatically requesting CoreLocation authentication.